requirements.txt
transformers==4.40.0
torch==2.2.1
python-chess==1.999
tqdm==4.66.4
fastapi==0.110.0
uvicorn==0.27.1
